requirements.txt
# Web framework
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0

# API clients
aiohttp>=3.8.0
ccxt>=4.0.0
websockets==12.0
requests==2.31.0

# Utilities
pydantic==2.4.2
